+++ StringDictionaryFMINDEX.h
@@ -75,7 +75,7 @@ public:
   @param strLen: the string length.
   @returns the ID (or NORESULT if it is not in the bucket).
    */
-  uint locate(uchar *str, uint strLen);
+  unsigned long locate(uchar *str, uint strLen);
 
   /** Obtains the string associated with the given ID.
   @param id: the ID to be extracted.
