+++ StringDictionaryRPHTFC.h
@@ -89,7 +89,7 @@ public:
       @param strLen: the string length.
       @returns the ID (or NORESULT if it is not in the bucket).
   */
-  uint locate(uchar *str, uint str_length);
+  unsigned long locate(uchar *str, uint str_length);
 
   /** Obtains the string associated with the given ID.
       @param id: the ID to be extracted.
